create new conftest files
