+++ tests/api/users/test_fetch_session_token.py
@@ -8,7 +8,7 @@ from boox.api.users import UsersApi
 from boox.core import Boox
 from boox.models.enums import BooxUrl
 from boox.models.users import DataToken, FetchTokenRequest, FetchTokenResponse
-from tests.api.conftest import FakeFetchTokenResponse
+from tests.api.users.conftest import FakeFetchTokenResponse
 from tests.api.utils import E2EConfig
 from tests.conftest import e2e
 
@@ -18,10 +18,10 @@ from tests.conftest import e2e
 def test_fetch_session_token_calls_post_and_parses_response(
     mocker: MockerFixture,
     faker: Faker,
-    fetch_token_response: FakeFetchTokenResponse,
+    fake_fetch_token_response: FakeFetchTokenResponse,
 ):
     mocked_response = mocker.Mock()
-    mocked_response.json.return_value = fetch_token_response.build().model_dump()
+    mocked_response.json.return_value = fake_fetch_token_response.build().model_dump()
     api = UsersApi(session=mocker.Mock())
     api._post = mocker.Mock(return_value=mocked_response)
 
@@ -36,12 +36,12 @@ def test_fetch_session_token_calls_post_and_parses_response(
 def test_users_api_fetch_session_token_integration(
     mocker: MockerFixture,
     faker: Faker,
-    fetch_token_response: FakeFetchTokenResponse,
+    fake_fetch_token_response: FakeFetchTokenResponse,
     mocked_client: mock.Mock,
     url: BooxUrl,
 ):
     mocked_response = mocker.Mock()
-    mocked_response.json.return_value = fetch_token_response.build().model_dump()
+    mocked_response.json.return_value = fake_fetch_token_response.build().model_dump()
     mocked_response.raise_for_status.return_value = mocked_response
     mocked_client.post.return_value = mocked_response
