+++ tests/api/users/test_get_user_info.py
@@ -9,7 +9,7 @@ from boox.api.users import UsersApi
 from boox.core import Boox
 from boox.models.enums import BooxUrl
 from boox.models.users import DataUser, UserInfoResponse
-from tests.api.conftest import FakeUserInfoResponse
+from tests.api.users.conftest import FakeUserInfoResponse
 from tests.api.utils import E2EConfig
 from tests.conftest import e2e
 
@@ -18,10 +18,10 @@ from tests.conftest import e2e
 
 def test_get_user_info_calls_get_and_parses_response(
     mocker: MockerFixture,
-    user_info_response: FakeUserInfoResponse,
+    fake_user_info_response: FakeUserInfoResponse,
 ):
     mocked_response = mocker.Mock()
-    mocked_response.json.return_value = user_info_response.build().model_dump()
+    mocked_response.json.return_value = fake_user_info_response.build().model_dump()
     api = UsersApi(session=mocker.Mock())
     api._get = mocker.Mock(return_value=mocked_response)
 
@@ -36,12 +36,12 @@ def test_get_user_info_calls_get_and_parses_response(
 def test_users_api_get_user_info_integration(
     mocker: MockerFixture,
     faker: Faker,
-    user_info_response: FakeUserInfoResponse,
+    fake_user_info_response: FakeUserInfoResponse,
     mocked_client: mock.Mock,
     url: BooxUrl,
 ):
     mocked_response = mocker.Mock()
-    mocked_response.json.return_value = user_info_response.build().model_dump()
+    mocked_response.json.return_value = fake_user_info_response.build().model_dump()
     mocked_response.raise_for_status.return_value = mocked_response
     mocked_client.get.return_value = mocked_response
